allow iframe title to be set. defaults to label
